WebRTC: Remove webrtc_perf_tests from trybots and Debug commit bots.
The webrtc_perf_tests test has grown to become a major bottleneck for
our CQ trybots since it's execution time has slowly increased over time.
It currently consumes 30-50% of the total build time for the trybots that
run tests on Android devices. It's even worse on Mac, Windows and Linux:
around 70% of the cycle time is consumed.

Given that failures only caught by this test is rare, it makes sense to
remove it from all bots that don't produce perf data (i.e. Release commit
bots).

I also added perf tracking for the not-yet-deployed Android devices that
currently runs in client.webrtc.fyi.

This is one step on the way to move all perf tests into a separate waterfall.

BUG=586238

Review-Url: https://codereview.chromium.org/2031673003