compile.py: Always default to ninja on mac; everything should be on ninja by now
This was still here 'cause I thought that the ios bots used that script, but turns out
they don't, so I could've removed this long ago.

If this affects your bot (it likely isn't a Chromium bot), please let me know.
Reverting this is fine, but the Right Fix is probably to pass an explicit
--build-tool parameter to compile.py for your bot.

BUG=294387,336575

Review-Url: https://codereview.chromium.org/2049893002