Remove explicit timeouts from win asan lkgr bots.
These bots used to run a VM with a custom clang bot.
Nowawadays, they use a baremetal builder with prebuilt clang, which means
they can use goma, which as of yesterday they do.  They now cycle in about
30 minutes, and timeout is how long a bot can be without any new output.
The default timeout is currently 40 minutes, so these bots easily complete
a whole build in the time a bot can be without new stdout.

BUG=none

Review-Url: https://codereview.chromium.org/2052003005