LogDog: Selectively use bootstrap success file.
Implement bootstrap success file support. Classicly, LogDog
Annotee/Butler will return a magic error code if they fail before
running their actual target, indicating that there was an error with the
bootstrap setup. This is not ideal, though, since return codes are a
fairly weak IPC channel and, additionally, both of these applications
forward arbitrary return codes from their target.

Support for Butler/Annotee emitting a "success" result file was recently
added. Now, enable waterfall configurations to selectively use this
support, and implement it. When enabled, the successful execution of the
bootstrapped process will identified by the existence of the result
file, and the return code from the process will be read from the file.

This is emitted as an optionally-enabled feature. If everything works, the
intention is to make this the default (and only) error detection method and
deprecate the magic return code values.

BUG=chromium:550673
TEST=unittest

Review-Url: https://codereview.chromium.org/2055403002