compile.py: Remove XcodebuildFilter
The filter is a great idea, but it broke when Xcode got newer than 3.2.
No bots use an Xcode that old, and we were never able to make it work with
newer Xcodes.  Since this hasn't been used anywhere in years for this reason,
and since no Chromium bots use xcodebuild anymore, delete this code.

BUG=92221,260989

Review-Url: https://codereview.chromium.org/2062823002