compile.py/mac: Stop emulating xcode flags in default build tool branch
If no --build-tool flag is passed explicitly, compile.py defaults to ninja.
For historical reasons, it also tried to parse -project and -target flags that
some bot configs used to pass (since they assumed that the default build tool
was Xcode on mac).  These bot configs are as far as I know long gone, so remove
that logic.

While here, also add a reference to the clobber bug to a TODO.

No intended behavior change.  If this does break some bot, then please revert
and tell me about it, but I'd be surprised by that.

If your bot uses Xcode, it should pass --build-tool=xcode (but it would've
already been broken before this change).

BUG=294387

Review-Url: https://codereview.chromium.org/2059243002