Conditionally json-parsing buildbucket property and its contents.
We have been getting an exception where one of the .get() calls in the
chain fails due to it being called on a 'unicode' rather than a dict.

We expected that reading dicts from properties wouldn't require parsing,
but it seems it does. This change is meant as a failsafe for that.

R=prasadv,nodir,martiniss
BUG=622044

Review-Url: https://codereview.chromium.org/2086663005