WebRTC: Remove get_version step from WebRTC bots.
WebRTC's Android bots are using the same toolchain as Android
for test execution on Android. Recent changes made the chromium_android
recipe module depend on the chrome/VERSION file in the checkout, which
doesn't exist. Workaround this by providing a way to skip the version
check and apply it to WebRTC's Android testers.

BUG=622606

Review-Url: https://codereview.chromium.org/2086333004