Rename Mac/Win GN builders to GYP builders.
Now that the main Mac and Win builders are on GN, the builders
with "GN" in the names are redundant and confusing. This CL renames
them and the corresponding trybots to "GYP" to provide some GYP
coverage until we drop the GYP build completely.

There are matching trybots for the builders, but they are not
part of the commit queue.

- chromium.mac
  - "Mac GN" -> "Mac GYP"
  - "Mac GN (dbg)" -> "Mac GYP (dbg)"
- chromium.win
  - "Win8 Aura" -> "Win8 GYP" (Aura has been default for a long time)
  - "Win8 GN (dbg) -> "Win8 GYP (dbg)"
- tryserver.chromium.mac
  - mac_chromium_rel_ng -> mac_chromium_gyp_rel
  - mac_chromium_dbg_ng -> mac_chromium_gyp_dbg
- tryserver.chromium.win
  - win8_chromium_ng -> win8_chromium_gyp_rel
  - win8_chromium_gn_dbg -> win8_chromium_gyp_dbg

R=rsesek@chromium.org
BUG=618468, 605318

Review-Url: https://codereview.chromium.org/2098943002