Run layout tests on Site Isolation Win FYI bot.
This CL modifies the bot's recipe to also run the full set of layout
tests with the --isolate-sites-for-testing=*.is flag.  This gives us
coverage for paths that are turned on by --isolate-extensions (we
can't use the actual --isolate-extensions flag, since content/ doesn't
know about it).  FlagExpectations/isolate-extensions will serve as the
disabled test list (similarly to site-per-process), and will be
uploaded in a separate CL.

BUG=477150, 532666

Review-Url: https://codereview.chromium.org/2100543002