Remove Windows/AMD from chromium.gpu and win_chromium_rel_ng.
The Chromium project has grown to the point where a significant number
of machines are required in order to keep up with the load on the main
commit queue. It's infeasible to expand all of the GPU types
simultaneously so this one is being moved to the chromium.gpu.fyi
waterfall. It will still be run for ANGLE CLs and on the optional GPU
tryservers.

BUG=624156

Review-Url: https://codereview.chromium.org/2108803003