archive: pass --src-dir to extract_build.py (part 1/2)
For now only enabled for 'Linux remote_run Tester'.

We need to extract to the checkout - some tests
require that to work, e.g. to get paths for test data.

For non-remote_run builders extract_build was able
to correctly guess the directory. remote_run uses a different
directory layout and so we need the flag.

BUG=chromium:585401, chromium:593999

Review-Url: https://codereview.chromium.org/2107173003