V8: Get build environment from MB
Before MB, the build environment was hard-coded by
configurations on the infra side. After moving to MB, the
build environment is set dynamically in a recipe step.

The build environment is written to a property and passed to
testers to be presented to the user on build failures. This
CL makes that work with MB, both if it's a gyp or gn
environment.

BUG=605516,616035

Review-Url: https://codereview.chromium.org/2127573003