Allowing zip_build to use a file name given in build_url.
In bisect and tryjobs we want to control from the recipe what the
filename of the uploaded zip will be. With this change we piggyback on
the build_url parameter to optionally pass such filename, and add a flag
to determine whether to interpret this url as containing such filname.

R=dtu,aiolos

Review-Url: https://codereview.chromium.org/2112413002