Fix configuration of goma in iOS recipes.
The way we were configuring the chromium module so that goma
would work in compiles didn't work correctly after we switched
the location of goma on the machine; the 'goma' config no longer
gives the right directory, and the `ensure_goma` step wasn't
being called reliably and consistently.

This CL reworks things so that we call ensure_goma consistently
when needed during the read_build_config() step of the ios
recipes, rather than during the build() step.

This CL also fixes an issue where if the GYP_DEFINES were specified
as a list of strings in the build config rather than a dict, things
wouldn't work right. We need this fix so that we can use goma
even with GYP, in a way that is consistent between GYP and GN.

R=smut@chromium.org
BUG=626033

Review-Url: https://codereview.chromium.org/2133483002