Add slave_lookup.py script.
This script dumps JSON configuration information for the set of slaves
read from either command-line or STDIN. It can be used by other scripts
or by troopers attempting to get information on a slave.

It's better than `botmap.py | grep ...` for this, since it shows more
fields and is also more accessible to other tooling and provides an
immediate response to the user.

TBR=stip@chromium.org
BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2127333003