V8: Fix tmp dir for gn comparison on windows.
This is working around http://crbug.com/361343 . GN doesn't
work if the output folder and the checkout are on different
drives on windows (on buildbot C: end E:).

This now uses a peek-gn output folder side-by-side to the
real output folder. Creation and removal of this folder are
now handled by the v8 api.

BUG=chromium:474921

Review-Url: https://codereview.chromium.org/2136033002