V8: Prevent v8 rolls while chromium checkout is inconsistent
This makes sure that the v8 versions from the pending ref
from the gerrit api and from the real ref from the local git
checkout are consistent.

If they are not, the roll would result in a patch failure
later. Inconsistencies stem from slow gnumbd service or
git backend inconsistency.

Review-Url: https://codereview.chromium.org/2140113002