recipes/win: Stop explicitly disabling precompiled headers in recipes in goma builds.
This isn't necessary: Both in gyp and gn builds, the .gyp and .gn files that
add the precompiled header flags know to only do this if goma isn't enabled.
This has to be done in gyp / gn files, so that if devs enable goma they only
need to set use_goma=true and don't have to set this other flag as well. So
rely on this in the recipe too.  (What the recipe does only has an effect in
gyp builds anyways, and gyp's going away.  And even then, it only has an effect
on bots not using mb, while virtually all bots should be on mb by now.)

BUG=none, but it's preparation for 627903

Review-Url: https://codereview.chromium.org/2151583002