buildbucket: support unique change urls
historically buildbucket was used for tryservers where changes are CLs
and a unique change id must be provided to correctly lookup a row in the
buildbot changes table. If unique id is not specified, a row in changes
table is added each time.

master.chromium.perf uses buildbucket differently. It polls chromium
changes from gitiles, compiles chromium and triggers testers on fyi. The
changes in the triggered builds are same valid chromium commits polled
from gitiles, as opposed to CLs in tryserver.

In this scenario a commit URL identifies a change, so use it to lookup a row in
the changes table

It also fixes changestore tests

R=dnj@chromium.org
BUG=626019

Review-Url: https://codereview.chromium.org/2157703002