Re-land https://codereview.chromium.org/2108553002/
Revert buildbot 361b81b43fc7793b66facd4bac7abe0adb30f3d5.

This commit was only added in the 0.8.4 branch, and was never made to master.

The bug only happens when starting buildbot with "make start", but we start it
directly with twistd in Chrome: http://trac.buildbot.net/ticket/1992

BUG=620572

Review-Url: https://codereview.chromium.org/2165433002