Remove 2GB bot. Rename FYI bots on perf dashboard.
The 2GB bot has never worked and we don't think we're going to be able to make it work in the near future (until isolated benchmarks). So, remove it. Then rename the "Low-End 4GB" bot to just "Low-End"

Rename the perf IDs of the FYI bots to match their bot names.

BUG=629560

Review-Url: https://codereview.chromium.org/2157343003