Clean up steps in Chromedriver recipe.
The tempfile recipe module creates a few messy looking steps and
the steps on the Chromedriver bot look very confusing and
unorganized. Adding nested steps for the test runs to clean things up.

Additionally, changed the name of server_log to make it unique for
every test run. It is currently always getting uploaded with the name
server_log and just overwriting itself.

BUG=627633

Review-Url: https://codereview.chromium.org/2164043002