Skip assigning provision profiles in the dummy test project.
The test project is trying to build and find a provisioning profile, but we
don't want the project to do anything with that.  The test apps are
signed while they are built.  Make the test project set the provisioning
profiles automatically to ignore it in case of the missing of provisions.

BUG=

Review-Url: https://codereview.chromium.org/2174713002