Add Clang Android CPU perf bots.
Clang is the future for the Android NDK, and we expect it to perform sometimes
signficantly differently to GCC.  As far as I can tell, the current matrix is:
  - ARMv7, GCC:   fine
  - ARMv8, GCC:   garbage
  - ARMv7, Clang: garbage   (this is a surprise)
  - ARMv8, Clang: great
Didn't bother comparing x86... I'm sure both are just great.

Am I right in remembering all we need to do is add bots with Clang in the name?

BUG=

Review-Url: https://codereview.chromium.org/2177953002