Make chromium_trybots always run layout tests on tryserver.blink bots.
Note how this makes it so that use_{skia,v8}_patch_on_blink_trybot
simulation expectation files include more tests. The webkit_tests
was in the simulation expectation files for the other blink try bots
because "suppress_analyze" was in the test, but that wasn't in the
step for use_{skia,v8}_patch_on_blink_trybot.

Pros:
 - This makes tryserver.blink bots behave similarly to the *_chromium_rel_ng builders

Cons
 - This would not make the new
linux_layout_tests_slimming_paint_v2 bot work as it should.
 - This results in running some blink-related tests that may not be necessary depending on the CL

This is one alternative to http://crrev.com/2190163003.

BUG=631014

Review-Url: https://codereview.chromium.org/2194553002