WebRTC: Enable MB for Linux bots and remove old way to run tests on GN bots.
This enables MB on our Linux bots except the libfuzzer bots
(configs were added in https://codereview.webrtc.org/2223893002/).

It also reverts https://codereview.chromium.org/2057453002/,
which is fine since when we have MB in place, users will be
able to use GN instead of GYP on any of the bots to build
and then run the already configured tests.

BUG=589510

Review-Url: https://codereview.chromium.org/2225353002