Gsutil ls misses should not be treated as failures by bisect.
The purpose of this is to declutter bisect output. The gsutil
module calls a wrapper that contains retry logic to avoid
cases of flake.

Actual failures of gsutil remain a possibility, and this
change would make it a little tricky to diagnose since we
won't have purple steps, but in this case all that will
happen is that builds will either time out or the download
and unzip step would fail.

Google Storage going down has not as of yet been a problem
for bisect.

R=sullivan,dtu@chromium.org
BUG=634045

Review-Url: https://codereview.chromium.org/2209833002