Update floating builder logic, add to "chromiumos"
The floating builder logic had several flaws, notably:
- It was broken: adding more than one (see crbug.com/600479).
- It was too complex. It was written generically with the idea that other
  waterfalls may want to use it (hah!)

This updates the logic to clearly have a primary and floating slave set
rather than a generic "strata" concept, and to specifically prefer them in a
predefined order. This also implements floating builders for PFQs (in builder
config), ups the number of floating builders from 1 to 2 for paladins, and
enrolls the floating builder logic for the "chromiumos" waterfall paladins.

BUG=chromium:637166
TEST=local

Review-Url: https://codereview.chromium.org/2250443002