Floating builder logic waits at startup.
Currently, the floating builder logic considered slaves that haven't
ever signed in as last seen a long time ago (epoch). This means that if
a master starts up, has a task, and a floating builder connects before a
primary, the floating builder will be immediately given the task, since
the primary is outside of the grace period.

Update this logic: now all slaves are considered last seen when the
floating function was created. This will cause the floating builder to
wait a full grace period before giving builds to floating slaves, even
if it just started up.

BUG=chromium:637166
TEST=None

Review-Url: https://codereview.chromium.org/2255643002