[perf] Reboot builders between builds.
We now have enough builder capacity (and building takes so much longer than
rebooting) that I'm comfortable with rebooting the builders between builds.
Not an urgent change as we don't have any current known issues with this,
so I won't schedule a master restart. Mostly, this is just in case.

BUG=628679

Review-Url: https://codereview.chromium.org/2257253002