Skia: add an NVIDIA Shield CPU perf bot.
I'd be happy for this bot to be anything we've got that's ARMv8 except the
Nexus 9.  The Nexus9 has a rather unusual CPU, while the Shield's is much more
representative of the bulk of ARMv8 chips.  I chose the Shield only because it
seems we have the most of them right now.

BUG=

Review-Url: https://codereview.chromium.org/2264603002