Created a line of summary and enabled suite name onclick function.
In suite result table, added a summary line to show the number of all
tests run, all tests failed etc.

If a suite name is clicked, only tests inside that suite, will be shown
in testcase table. If 'TOTAL' is clicked in the summary line, tests in
all suites will be shown in testcase table.

BUG=

Review-Url: https://codereview.chromium.org/2252603002