Use files in path['checkout'] not resource.
To make it easy for people to update compare_build_artifacts.py
and deterministic_build_blacklist.json, we have already imported
them in chormium src.
remove_build_metadata.py was also imported in chromium src.

This change is for making isolate APIs to use files in chromium src
not in resources.

BUG=314403

Review-Url: https://codereview.chromium.org/2260733002