Reland of 'use buildbucket to assess build failure'.
Original CL: https://codereview.chromium.org/2274763002/
In the reverted patch, I had mistakenly assumed that the response of buildbucket.get was inside a list, as the value for the 'results' key of a dict. I probably copied that structure from somewhere and mocked the response to match, hence the sim test did not catch this error.

BUG=638599
R=prasadv,sullivan

Review-Url: https://codereview.chromium.org/2276273003