Do not exclude the last revision in the range returned by gitiles.
Originally we excluded the last revision in the range, because that
operation included the revision given as the end. When expanding the
revision range the 'bad' revision is the end revision, and we don't want
to instantiate it again, since we already do that when starting the
bisect jobs. When reusing this code for expanding rolls, we ended up
also discarding the end revision, which makes for one-revision rolls to
break and also excludes the last revision in the roll.

BUG=641133
R=prasadv,dtu,sullivan

Review-Url: https://codereview.chromium.org/2282563004