Increase timeout for determining initial confidence from 2 hours to 20 hours.
Some benchmarks take up to 40 minutes to run. 2 hours only gets you 4
iterations. That's not enough to establish confidence for many regressions. If
you want to get up to 15 iterations of both the first and last commits, that's
20 hours. A build times out after 24 hours, so 20 seems like a good number with
some headroom.

Unfortunately, if it takes that long to establish confidence, the bisect run
will probably time out after 24 hours, anyway. You could argue that we're
wasting bisect resources doing this, but having a timeout would be clearer to
the user than having a no-confidence result.

Long term fix would be to implement --story-filter for benchmarks that we can
divide up, to reduce the run time:
https://github.com/catapult-project/catapult/issues/1811
Also swarming will remove that 24-hour timeout restriction, I believe.

BUG=640509

Review-Url: https://codereview.chromium.org/2273173002