Reland 'Add buildbot metadata to compiler_proxy log'
We'd like to analyze compiler proxy log. However, buildbot metadata is
missing, so sometimes hard to determine in which builder the compiler
proxy log is generated.

Append buildbot metadata to compiler_proxy log like ninja log.

The difference from the previous patch is to add '--' to gsutil.py.
Without '--', gsutil.py consumes options which is expected to pass
to gsutil itself.

Review-Url: https://codereview.chromium.org/2311123002