3rdtry: 'Add buildbot metadata to compiler_proxy log'
We'd like to analyze compiler proxy log. However, buildbot metadata is
missing, so sometimes hard to determine in which builder the compiler
proxy log is generated.

Append buildbot metadata to compiler_proxy log like ninja log.

The difference from the first try is to add '--' to gsutil.py.
Without '--', gsutil.py consumes options which is expected to pass
to gsutil itself.

The difference from the second try is custom metadata starts with
x-goog-meta-. Custom metadata must start with x-goog-meta
according to
https://cloud.google.com/storage/docs/gsutil/addlhelp/WorkingWithObjectMetadata#custom-metadata

Review-Url: https://codereview.chromium.org/2312743004