recipe_modules/chromite: Use "build_type".
Specialized config-specific parameters were previously being pushed from
the master to the Chromite recipe, creating an information channel and
dependency between the two.

Chromite specifies all of the information that is needed in its
configuration JSON, and that file is available to the recipe during
execution. Load the build type from the configuration JSON instead of
having it be pushed by the master.

BUG=chromium:627996
TEST=expectations

Review-Url: https://codereview.chromium.org/2325913002