Corrected test results presentation for android gtests; Made method showTestsOfOneSuiteOnly faster.
In gtest, suite name and test name is separated by '.', instead of '#' as
in instrumentation test. This cl fixes this and thus test results
presentation is able to recognize suite name and test name of gtest correctly.

When there are too many tests, showTestsOfOneSuiteOnly took too long,
around 2~3 seconds. In this cl, showTestsOfOneSuiteOnly is able to
run faster, by calling simpler built-in methods.

BUG=605572

Review-Url: https://codereview.chromium.org/2355713003