Build PDFium out_dir
The bots can build multiple configurations for PDFium which may mean we flip
between things like XFA and non-XFA. This results in a lot of un-needed
recompilation.

This Cl makes an output directory for each of the configurations that we build
instead of always using out/Release or out/Debug.

BUG=pdfium:600

Review-Url: https://codereview.chromium.org/2357503003