WebRTC: Reshuffle Mac VMs to prevent errors due to mixed Xcode versions.
In crbug.com/649150 we're upgrading machines to Xcode 8, but since not all
machines are done yet and the find_xcode step fails if the found version
doesn't match the one configured in the JSON configs in
https://chromium.googlesource.com/external/webrtc/+/master/webrtc/build/ios/client.webrtc/
the Xcode 8 machines needs to be taken out to prevent breaking client.webrtc.

One VM was swapped with client.webrtc.fyi since that one was also still running
Xcode 7.

BUG=649150

Review-Url: https://codereview.chromium.org/2364873002