WebRTC: Move Android swarming testers to separate VMs
We had issues with tester builders running on the same machine as a
builder picking up triggered builds at the same time as the VM is being
shut down. Moving them to separate machines for now. This may not be
optimally balanced but we're planning to change this shortly anyway.

BUG=583318
TBR=ehmaldonado@webrtc.org

Review-Url: https://codereview.chromium.org/2399773003