Enable saving the same file in multiple archives
Earlier I submitted a CL to archive content shell and I ran into an issue
where the current FILES.cfg parser doesn't allow listing the same file
for multiple archives.

I opted to allow multiple entries of the same file in FILES.cfg rather than
making the 'archive' property a list because this allows more configurability
(e.g. for icudtl.dat, we only need to archive it for content shell for dev builds
but for the main archive we archive it for dev and official builds).

BUG=654486

Review-Url: https://codereview.chromium.org/2402423002