Log more when loading shard output.json fails
When we use swarming to run steps for build bots, and
swarming fails to load the output.json for and shard, the
whole step is marked as an infra failure.

Sometimes, the shard completes successfully, but swarming
can't load the output.json for whatever reason. We're not
sure why this is happening, and hopefully this will help.

BUG=653728

Review-Url: https://codereview.chromium.org/2413553002