Remove setting of options from factory_properties in archive_layout_test_results.py.
Reason: There is no longer a --factory-properties option for this script, and --gs-bucket is passed in as a command line parameter.

I don't think we're actually using gs_acl, and in actual runs of this script, no --gs-acl is passed in. So as a follow-up, I think I may be able to remove the --gs-acl parameter from this script.

Review-Url: https://codereview.chromium.org/2416453005