Enable merging buildbot builds dependent on queue length
This enhances custom mergeRequests functions with additional
parameters for the number of unclaimed builds in the queue
and the number of already merged builds.

This allows writing more flexible merge strategies, e.g.:
- Avoid huge blame lists during peak hours that stem from
the standard merge-all strategy.
- Avoid huge backlog on bots using the merge-none strategy.

This also adds an example FYI bot from V8 to test the new
feature.

BUG=654749
TEST=Tested on a local master stressed with fake commits.

Review-Url: https://codereview.chromium.org/2415703004