Remove use of "use_mirror" recipe property.
The "use_mirror" property is set by the recipe engine when in a
production environment to signal to the "gclient" recipe module that the SVN
GOLO mirror should be used instead of the default SVN URL. Since we're
not on SVN, this is now meaningless. It will be deprecated in recipe
engine, so remove references.

The "perf_dashboard" recipe has been using this property as a test to see if
it is running on a production system. This test has been replaced with
the identical logic that the recipe engine was using to set
"use_mirror". A build property has been added to indicate production
runs, and a note has been added to indicate that the perf dashboard bot
should explicitly specify this as a property rather than infer it from
the environment.

BUG=chromium:656741
TEST=expectations
R=iannucci@chromium.org, martiniss@chromium.org

Review-Url: https://codereview.chromium.org/2427763002