Archive retry summary for layout tests along with layout test results.
Motivation: For `webkit-patch rebaseline-cl`, I want to be able to
easily get a list of layout tests that failed with the patch in a try
job but not without the patch. This information is available in the retry
summary step after running without the patch. But, the retry summary is
not currently stored anywhere that is convenient to fetch.

If the retry summary is stored along with the results for the first
run of the layout tests, then it will be simple to fetch and use when
deciding what layout tests to rebasline.

BUG=642980

Review-Url: https://codereview.chromium.org/2421733002