DevTools: add closure compilation step
This runs closure compiler for DevTools front-end code. Although there
is a closure compiler for chromium, DevTools has a modified closure compiler
and uses it in a different way to analyze our modules' dependency graph.

I've disabled the npm/eslint steps until we get open source compliance approval
(I've gotten most of the necessary approvals and waiting for one more).

This assumes there's JVM 7 or above on the buildbot slave.

BUG=none

Review-Url: https://chromiumcodereview.appspot.com/2433493002