Change archive_retry_summary step to happen also when there are new failures.
In http://crrev.com/2421733002, I added a step after
"webkit_tests (retry_summary)" that calls archive_layout_test_retry_summary.py
which archives the retry summary JSON along with the layout tests results
in Google Storage.

But, this step was added in a function after a try/finally block, where
if there are failures that didn't occur without the patch, a StepFailure
exception is raised inside the try block (I think) -- so, when there is
a StepFailure raised, the retry results are not archived, which is not
what I wanted.

This CL just changes it so that the archive retry results step is inside
the finally block so that the step should always appear after
"webkit_tests (retry summary)".

BUG=642980

Review-Url: https://chromiumcodereview.appspot.com/2434353002