Improve slave-preference function
The current slave-preference function has the following
flaw:

When the system is under medium load, preferred slaves
are often occupied, so that each new build starts taking
slaves preferring other builders. This in turn leads to more
slaves not being available for their preferred builders.

The new implementation has the following strategy:

Let slaves(builder) be the number of available slaves
preferring that builder. For builder A and available
slaves S, choose s in S such that (order is preference):
1. s prefers A, or
2. s prefers nothing (an optional fall-over pool), or
3. s prefers some B != A and there's no other builder C != B
such that slaves(C) > slaves(B)

BUG=658189

Review-Url: https://codereview.chromium.org/2427413005