Change all remote run builders to use tools/build.
Recipe rolling is stuck too often, and is causing CLs in build to not take
effect for a day, which is bad. For the time being, we're going to use remote
run on the recipes in tools build, so that the roller is not in a critical
path (except for build_internal, but that's a much smaller volume of CLs).

BUG=584079

Review-Url: https://codereview.chromium.org/2445193002