Avoid unnecessary file churn by "git reset --hard" step.
Some of the build step create hard-links to files in the checked out
repository which change the file ctime timestamp even. In turn, this
cause "git reset --hard" to consider the file as modified due to a
cache mismatch between its index and the checkout.

As a workaround, add a "git update-index --refresh" call to force a
refresh of the git index, which will update the ctime stored in the
cache, thus avoiding the unnecessary recreation of the file. This is
necessary as otherwise ninja would consider the output of the
hardlink step as stale and would re-run the corresponding step
thus creating a vicious cycle.

BUG=642711

Review-Url: https://codereview.chromium.org/2444403003