WebRTC: Add new Linux (more configs) bots.
This new bot runs GN with a minimal amount of configurations set,
then compiles. This will give us coverage for the cases where
dcheck_always_on=false and rtc_use_h264=false.

It also re-runs GN+compile with:
* rtc_enable_intelligibility_enhancer=true
* rtc_enable_protobuf=false
* rtc_include_opus=false

If we want to test more configurations like this, we can just add
them to this bot in the future.

BUG=652197,611054

Review-Url: https://codereview.chromium.org/2460493002