Setting a maximum number of test runs per revision.
The chosen value is 20 runs per revision. The reasoning is that 20 is a
round number that should produce no less than 18 values (one per run, in
a test that is at most 10% flaky) which has in the past seemed
sufficient to decide that the values are unlikely to converge per MWU.

Also, adding test coverage and fixing a small bug highlighted by this
change. (See bisector.py one line change)

R=dtu@chromium.org
BUG=659906

Review-Url: https://codereview.chromium.org/2469503002