Add op args to remote run and annotated run.
https://codereview.chromium.org/2360083008 adds more exit information to the recipe
engine. This CL uses this new information to more correctly determine the exit
status of a recipe execution attempt. This will make it so that the
remote_run_result step is no longer colored purple when a build exits with an
exception, which allows for better flakiness detection logic.

Note that this CL should cause no behavior changes; everything is hidden behind
flags which aren't turned on anywhere.

BUG=646142

Review-Url: https://codereview.chromium.org/2380443002