WebRTC: Set the right path to clean_symlinks.py
The path calculated in https://codereview.chromium.org/2484803004
missed the fact that the checkout root was the dir that was passed
to the script. Update the cleanup_symlinks.py script to set the
right path and clarify that the directory argument to pass to it
is expected to be the checkout root (not the path to the src/ dir).
Also make the script not fail if the file is not found (a new bot don't have it, but
still has the checkout root).

BUG=663278
TBR=ehmaldonado@chromium.org

Review-Url: https://codereview.chromium.org/2482153002