WebRTC: Fallback for working_dir in symlink cleaning
The solution for the symlink cleaning in
https://codereview.chromium.org/2494663002/ doesn't seem to have the
desired effect (the checkout being used isn't found).
Change back to the previous solution but add a fallback for recipes
not using the remote_run way of setting up the directory
hierarchy on the bot machine.

BUG=663278
TBR=ehmaldonado@chromium.org
TESTED=
Ran scripts/slave/recipe_modules/webrtc/resources/cleanup_symlinks.py
with and without argument to ensure no syntax errors.

Review-Url: https://codereview.chromium.org/2491513003