Update waterfall templates to include Milo link.
Milo is currently ingesting all of the build data for all of the
waterfalls. Add a link to each waterfall with a parallel Milo view that
clicks through to Milo.

In order to do this, we add core "milo.js" and "milo.css" files to
"chromium"'s "public_html" folders. We also augment "layout.html" to
execute the "milo.js" injection function.

Every waterfall with its own "layout.html" template has been updated to
include the "milo.js" import and injection code.

Every waterfall with its own "public_html" directory has had the Milo
JavaScript, CSS, and image files symlinked to "chromium/" so they can
access the resources referenced by their "layout.html".

The injection function examines the current URL and decides if it maps
to a view Milo supports, namely a builder or build view. If it does, a
[LogDog] tab is injected into the header that links to the Milo
equivalent view.

The BuildBot base template is augmented to pass the master name to
"milo.js". This is the actual name used by Milo, as opposed to the
chromegw URL name which can occasionally vary. The modified template
parameter is protected with the "|default" filter function so masters
whose templates update but aren't restarted won't fail to render.

BUG=None
TEST=local
  - Started masters with affected "layout.html" pages locally, seem to
    be working!

R=estaab@chromium.org, iannucci@chromium.org

Review-Url: https://codereview.chromium.org/2491643003