Use devil's adb with each bisect iteration, also start/stop daemons.
In order to cleanly accomplish this change, custom android code has been moved out of the auto bisect loop and into two context managers. Desktop and other non-Android builds will use noop context managers instead.

This change cleans up a lot of nested logic in the bisect and makes it a little less commingled with calling code.

This change is only for the staging recipe. If shown to work, we can land https://codereview.chromium.org/2485763002/.

BUG=660852

Review-Url: https://codereview.chromium.org/2486923002