Temporary remove WATCHLISTS file.
It has only kjellander-cc@chromium.org in it, and this account is unknown to
Gerrit and Gerrit complains. See the bug for more details.

Removing just this account from the file makes it empty and presubmit complains
with "ERROR:root:WATCHLISTS not defined in ...", so removing the whole file for
now.

BUG=665640
TBR=agable@chromium.org

Change-Id: Ibfacff9c69bc4e8c260f8f2cea16834fe22d435f
Reviewed-on: https://chromium-review.googlesource.com/411922
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>