Fixed the class of the row of crashed testcase.
Previously the class of the row of crashed testcase was categorized as
'failure'. However we just realized that tombstones will only be
presented only if the class of the row is 'crashed'. Thus in this cl,
added 'crashed' into the class of the row of crashed testcase.

BUG=666298

Change-Id: Icf737ef0889a98f8769ede59404f0846aaa5d00e
Reviewed-on: https://chromium-review.googlesource.com/412283
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Zhiling Huang <hzl@google.com>