Remove duplicate testruns in test results presentation.
The same testcase can be run multiple times and all these testruns show
up in the json file. Previously test results presentation show all the
testruns of the same testcase, which is not what we desire.

In this cl, we remove duplicate testruns. And the result of later
testruns will replace the result of previous testruns.


BUG=666368

Change-Id: Icbd0017da9e1a87a7eedfb202dfaf103466261b8
Reviewed-on: https://chromium-review.googlesource.com/412386
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Zhiling Huang <hzl@google.com>